Split hotspot files by concern
